Create joystick_manager.cpp and joystick_manager.h
